+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Fixed
 * Disable "Save Changes" and "Reset Changes" buttons when no flaw changes have been made (`OSIDB-4973`)
+* Fix my issues flaw list filter to use email instead username (`OSIDB-5143`)
 
 ### Changed
 * Update error message on failed logging (`OSIDB-5011`)